+++ docs/conf.py
@@ -6,6 +6,8 @@
 # full list see the documentation:
 # http://www.sphinx-doc.org/en/master/config
 import re
+import os
+import sys
 
 
 def setup(app):
@@ -18,8 +20,7 @@ def setup(app):
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-import os
-import sys
+
 sys.path.insert(0, os.path.abspath('..'))
 
 autodoc_member_order = 'bysource'
@@ -55,6 +56,7 @@ extensions = [
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages',
     'sphinx.ext.napoleon',
+    'sphinx.ext.coverage',
     'sphinxcontrib.asyncio'
 ]